# Optimize memory consumption