tested with 

`blackfire run --ignore-exit-status php vendor/bin/phpunit tests/PHPStan/Analyser/NodeScopeResolverTest.php`

before this PR: (1.10.x@7c4ef6b2a)
```
I/O Wait      171ms
CPU Time   2min 13s
Memory        368MB
Wall Time  2min 13s
Network         n/a     n/a     n/a
SQL             n/a     n/a
```

after this PR:
```
Wall Time  2min 16s
I/O Wait      183ms
CPU Time   2min 16s
Memory        345MB
Network         n/a     n/a     n/a
SQL             n/a     n/a
```

<img width="304" alt="grafik" src="https://github.com/phpstan/phpstan-src/assets/120441/5f1f2a56-44aa-40a5-806c-a1458f06299e">